requirements.txt
ortools>=9.0.0
pycairo>=1.20.0
Pillow>=10.0.0
aggdraw>=1.3.18 